Level Update
Added the tileset we are going to use.
Added camera tools to have the camera follow the player.
Did animation for run, idle and jump (Jump animation is a bit wonky. Still need to fix that.)
Animation for crouch, attack and hit is done. Just needs to be set up and coded.
Added code to Controller script for events.